--- recipes/catch2/2.13.4/conandata.yml
@@ -1,4 +1,0 @@
-sources:
-  2.13.4:
-    url: https://github.com/catchorg/Catch2/archive/v2.13.4.tar.gz
-    sha256: e7eb70b3d0ac2ed7dcf14563ad808740c29e628edde99e973adad373a2b5e4df
