--- recipes/catch2/2.13.4/test_package/conanfile.py
@@ -1,17 +1,0 @@
-from conans import ConanFile, CMake, tools
-import os
-
-
-class TestPackageConan(ConanFile):
-    settings = "os", "compiler", "build_type", "arch"
-    generators = "cmake", "cmake_find_package"
-
-    def build(self):
-        cmake = CMake(self)
-        cmake.configure()
-        cmake.build()
-
-    def test(self):
-        if not tools.cross_building(self.settings):
-            self.run(os.path.join("bin", "test_package"), run_environment=True)
-            self.run(os.path.join("bin", "standalone"), run_environment=True)
