+++ recipes/catch2/2.x.x/test_package/CMakeLists.txt
@@ -7,5 +7,12 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 include(${CMAKE_BINARY_DIR}/conanbuildinfo.cmake)
 conan_basic_setup()
 
-add_executable(${PROJECT_NAME} 000-CatchMain.cpp 100-Fix-Section.cpp)
-target_link_libraries(${PROJECT_NAME} ${CONAN_LIBS})
+find_package(Catch2 REQUIRED)
+
+add_executable(test_package 000-CatchMain.cpp 100-Fix-Section.cpp)
+target_link_libraries(test_package PRIVATE Catch2::Catch2)
+
+if(Catch2_VERSION VERSION_GREATER_EQUAL 2.13.4)
+  add_executable(standalone 200-standalone.cpp)
+  target_link_libraries(standalone PRIVATE Catch2::Catch2WithMain)
+endif()
