+++ recipes/catch2/2.x.x/test_package/conanfile.py
@@ -4,7 +4,7 @@ import os
 
 class TestPackageConan(ConanFile):
     settings = "os", "compiler", "build_type", "arch"
-    generators = "cmake"
+    generators = "cmake", "cmake_find_package"
 
     def build(self):
         cmake = CMake(self)
@@ -13,5 +13,7 @@ class TestPackageConan(ConanFile):
 
     def test(self):
         if not tools.cross_building(self.settings):
-            bin_path = os.path.join("bin", "test_package")
-            self.run(bin_path, run_environment=True)
+            self.run(os.path.join("bin", "test_package"), run_environment=True)
+
+            if self.deps_cpp_info["catch2"].version >= "2.13.4":
+                self.run(os.path.join("bin", "standalone"), run_environment=True)
