+++ recipes/catch2/config.yml
@@ -26,4 +26,4 @@ versions:
   2.13.3:
     folder: 2.x.x
   2.13.4:
-    folder: 2.13.4
+    folder: 2.x.x
